client: Implicitly use default port if not specified.
Running the IRC client with the -s option without the -p option
would result in it just starting an idle client (without connecting
to the server), because of the server && port condition, with
port defaulting to 0. The IRC library does take care of defaulting
the port, but due to the way this logic is written, we now default
the port in the client itself up front so that the port is truly optional.