client: Don't abort if unable to write to log file.
If the LIRC client is launched twice, it may try to write to the same log file,
so subsequent invocations currently fail as a result. We now continue
if unable to write to the log file so multiple instances can be launched.

A possible future improvement here is to allow specifying the log file
to use as an argument.